completted Order by exercise and Limit exercise

+++ albums_migration.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS albums (
     artist VARCHAR(50) NOT NULL,
     album VARCHAR(100) NOT NULL,
     release_date YEAR (4) NOT NULL,
-    sales_in_Millions DOUBLE(3,1) UNSIGNED NULL,
+    sales_in_Millions DOUBLE(5,1) UNSIGNED NULL,
     genre VARCHAR(50) NOT NULL,
     PRIMARY KEY (id)
 );
